pyproject.toml
[tool.poetry]
name = "personalsite"
version = "0.1.0"
description = ""
authors = ["Mohan Dhar <mohan@albert.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
django-heroku = "^0.3.1"
python-dotenv = "^0.19.0"
django-phonenumber-field = "^5.2.0"
phonenumbers = "^8.12.28"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
